image_diff/requirements.txt
transformers
Pillow
torch
scikit-learn
